Move setup and teardown into a BasicTestCase class
